Extrae la pantalla de configuracion
